+++ ddserver/resources/doc/schema.upgrade.sql
@@ -16,10 +16,17 @@
 -- along with ddserver.  If not, see <http://www.gnu.org/licenses/>.
 
 
-
 --
 -- Upgrading from version 0.2.x
 --
+
+ALTER TABLE `users`
+  ADD `yubico_id` INT NULL DEFAULT NULL AFTER `password` ,
+  ADD `yubico_key` VARCHAR( 32 ) NULL DEFAULT NULL AFTER `yubico_id` ;
+
+ALTER TABLE `hosts`
+  ADD `abuse` TEXT NULL DEFAULT NULL ;
+
 ALTER TABLE `hosts`
   ADD `address_v6` VARCHAR(39) NULL DEFAULT NULL ;