+++ CHANGES.md
@@ -6,6 +6,9 @@ New features:
 * Added a ddserver logo
 * Added support for python 3.3
 * Use the public require library insted of the out-dated buildin one
+* Added support for Yubikey OTP authentication
+* Allow administrators to disable hostnames (i.e., on abuse)
+* Use the powerdns remote backend, instead of the pipe backend
 
 Fixes: